requirements.txt
biopython==1.73
numpy==1.16.0
pandas==0.24.0
python-dateutil==2.7.5
pytz==2018.9
six==1.12.0
